MacOS build & apple clang warnings fixes (#122)
* [No Issue] Features incompatible with MacOS hided under defines

* [No Issue] Unused arguments commented out

Reason: To please clang warnings check

* [No Issue] Added cast to void* for printf arguments

Reason: To please clang warnings check, std requires arg to be void*, not char*

* [No Issue] Removed unused global variables

Reason: To please clang warnings check

* [No Issue] Implicit cast made explicit

Reason: To please clang warnings check

* [No Issue] Removed extra move of unique_ptr

Reason: To please clang warnings check.

* [No Issue] Added required include

Reason: build fails otherwise on MacOS

* [No Issue] Old fashion cast replaced with regular

Reason: To please clang warnings check

* [No Issue] aligned_allocate.ipp missprint fix

+++ benchmark/message/fixtures.hpp
@@ -27,7 +27,7 @@ namespace benchmark_messages {
         public:
             virtual void SetUp(const ::benchmark::State& state) final {
                 for (int i = 0; i < state.range(0); ++i) {
-                    name_ = i % 10;
+                    name_ = static_cast<uint64_t>(i) % 10;
                 }
             }
